refactor: PreviewPane types を分割し file-length linter をパスするように修正
